@@ -2,25 +2,34 @@
     "private": true,
     "scripts": {
         "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --disable-host-check --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "development": "npx mix",
+        "watch": "npx mix watch",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "npx mix -p"
     },
     "devDependencies": {
-        "axios": "^0.19",
+        "axios": "^0.21.1",
         "bootstrap": "^4.0.0",
-        "cross-env": "^7.0",
+        "cross-env": "^7.0.3",
         "jquery": "^3.2",
-        "laravel-mix": "^5.0.1",
+        "laravel-elixir-vue-2": "^0.3.0",
+        "laravel-mix": "^6.0.16",
         "lodash": "^4.17.13",
         "popper.js": "^1.12",
+        "postcss": "^8.1.14",
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.20.1",
         "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
+        "vue": "^2.6.12",
+        "vue-loader": "^15.9.5",
         "vue-template-compiler": "^2.6.10"
+    },
+    "dependencies": {
+        "bootstrap-vue": "^2.21.2",
+        "buefy": "^0.9.6",
+        "material-icons": "^0.6.1",
+        "npm-watch": "^0.9.0",
+        "vue-router": "^3.5.1",
+        "vuesax": "^4.0.1-alpha.25"
     }
 }
