@@ -6,5 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class Extra extends Model
 {
-    //
+    protected $fillable = [
+        'title', 'status',
+    ]; 
 }
