@@ -1,4 +1,5 @@
-const elixir = require('laravel-elixir');
+
+var elixir = require('laravel-elixir')
 
 require('laravel-elixir-vue-2');
 
